+++ gulpfile.js
@@ -1,12 +1,17 @@
+const { join, relative } = require('path')
+const { cwd } = require('process')
+const { Transform } = require('stream')
+const { fileURLToPath } = require('url')
+
 const autoprefixer = require('autoprefixer')
 const cssnano = require('cssnano')
 const gulp = require('gulp')
-const clean = require('gulp-clean')
 const postcss = require('gulp-postcss')
 const rename = require('gulp-rename')
 const gulpSass = require('gulp-sass')
 const terser = require('gulp-terser')
-const dartSass = require('sass')
+const PluginError = require('plugin-error')
+const dartSass = require('sass-embedded')
 const webpack = require('webpack-stream')
 
 const { version } = require('./package.json')
@@ -17,9 +22,14 @@ const { version } = require('./package.json')
  */
 require('./tasks/docs')
 
-/* Remove all compiled files */
-function cleanDist() {
-  return gulp.src('dist', { allowEmpty: true }).pipe(clean())
+/**
+ * Remove all compiled files
+ * @param {string[]} pattern - glob patterns or paths to clean
+ * @param {string[]} [ignore] - glob patterns or paths to ignore
+ */
+async function clean(pattern, ignore) {
+  const { deleteAsync } = await import('del')
+  await deleteAsync(pattern, { ignore })
 }
 
 /**
@@ -31,26 +41,66 @@ const sass = gulpSass(dartSass)
 /* Build the CSS from source */
 function compileCSS(done) {
   return gulp
-    .src(['packages/nhsuk.scss'])
-    .pipe(sass().on('error', done))
+    .src(['packages/nhsuk.scss'], {
+      sourcemaps: true
+    })
+    .pipe(
+      sass({
+        sourceMap: true,
+        sourceMapIncludeSources: true
+      }).on('error', (error) => {
+        done(
+          new PluginError('compileCSS', error.messageFormatted, {
+            showProperties: false
+          })
+        )
+      })
+    )
+    .pipe(
+      new Transform({
+        objectMode: true,
+
+        // Make source file:// paths relative
+        transform(file, enc, cb) {
+          if (file.sourceMap?.sources) {
+            file.sourceMap.sources = file.sourceMap.sources.map((path) =>
+              relative(join(cwd(), 'dist'), fileURLToPath(path))
+            )
+          }
+
+          cb(null, file)
+        }
+      })
+    )
     .pipe(postcss([autoprefixer()]))
-    .pipe(gulp.dest('dist/'))
+    .pipe(
+      gulp.dest('dist/', {
+        sourcemaps: '.'
+      })
+    )
 }
 
 /* Minify CSS and add a min.css suffix */
 function minifyCSS() {
   return gulp
-    .src([
-      'dist/*.css',
-      '!dist/*.min.css' // don't re-minify minified css
-    ])
+    .src(
+      [
+        'dist/*.css',
+        '!dist/*.min.css' // don't re-minify minified css
+      ],
+      { sourcemaps: true }
+    )
     .pipe(postcss([cssnano()]))
     .pipe(
       rename({
         suffix: `-${version}.min`
       })
     )
-    .pipe(gulp.dest('dist/'))
+    .pipe(
+      gulp.dest('dist/', {
+        sourcemaps: '.'
+      })
+    )
 }
 
 /**
@@ -58,11 +108,14 @@ function minifyCSS() {
  */
 
 /* Use Webpack to build and minify the NHS.UK components JS. */
-function webpackJS() {
+function webpackJS(done) {
   return gulp
-    .src('./packages/nhsuk.js')
+    .src('./packages/nhsuk.js', {
+      sourcemaps: true
+    })
     .pipe(
       webpack({
+        devtool: 'source-map',
         mode: 'production',
         module: {
           rules: [
@@ -80,24 +133,49 @@ function webpackJS() {
           minimize: false // minification is handled by terser
         },
         output: {
-          filename: 'nhsuk.js'
+          filename: 'nhsuk.js',
+
+          // Make source webpack:// paths relative
+          devtoolModuleFilenameTemplate(info) {
+            return relative(join(cwd(), 'dist'), info.absoluteResourcePath)
+          }
+        },
+        stats: {
+          colors: true,
+          errors: false
         },
         target: 'browserslist'
+      }).on('error', (error) => {
+        done(
+          new PluginError('webpackJS', error, {
+            showProperties: false
+          })
+        )
       })
     )
-    .pipe(gulp.dest('./dist'))
+    .pipe(
+      gulp.dest('./dist', {
+        sourcemaps: '.'
+      })
+    )
 }
 
 /* Minify the JS file for release */
 function minifyJS() {
   return gulp
-    .src([
-      'dist/*.js',
-      '!dist/*.min.js' // don't re-minify minified javascript
-    ])
+    .src(
+      [
+        'dist/*.js',
+        '!dist/*.min.js' // don't re-minify minified javascript
+      ],
+      { sourcemaps: true }
+    )
     .pipe(
       terser({
         format: { comments: false },
+        sourceMap: {
+          includeSources: true
+        },
 
         // Compatibility workarounds
         ecma: 5,
@@ -106,23 +184,14 @@ function minifyJS() {
     )
     .pipe(
       rename({
-        suffix: '.min'
+        suffix: `-${version}.min`
       })
     )
-    .pipe(gulp.dest('dist/'))
-}
-
-/* Version the JS file for release */
-function versionJS() {
-  return gulp
-    .src('dist/nhsuk.min.js')
     .pipe(
-      rename({
-        basename: `nhsuk-${version}`,
-        extname: '.min.js'
+      gulp.dest('dist/', {
+        sourcemaps: '.'
       })
     )
-    .pipe(gulp.dest('dist/'))
 }
 
 /**
@@ -144,28 +213,24 @@ function assets() {
 
 /* Copy JS files into their relevant folders */
 function jsFolder() {
-  return gulp
-    .src('dist/*.min.js', '!dist/js/nhsuk.min.js')
-    .pipe(clean())
-    .pipe(gulp.dest('dist/js/'))
+  return gulp.src('dist/*.min.{js,js.map}').pipe(gulp.dest('dist/js/'))
 }
 
 /* Copy CSS files into their relevant folders */
 
 function cssFolder() {
-  return gulp.src('dist/*.min.css').pipe(clean()).pipe(gulp.dest('dist/css/'))
+  return gulp.src('dist/*.min.{css,css.map}').pipe(gulp.dest('dist/css/'))
 }
 
 async function createZip() {
-  const zip = await import('gulp-zip')
+  const { default: zip } = await import('gulp-zip')
 
   return gulp
     .src(
       [
-        'dist/css/*.min.css',
-        'dist/js/*.min.js',
-        'dist/assets/**',
-        '!dist/js/nhsuk.min.js'
+        'dist/css/*.min.{css,css.map}',
+        'dist/js/*.min.{js,js.map}',
+        'dist/assets/**'
       ],
       {
         base: 'dist',
@@ -180,38 +245,36 @@ async function createZip() {
  * Development tasks
  */
 
-gulp.task('clean', cleanDist)
+gulp.task('clean', async () => {
+  return clean(['dist/**/*'])
+})
 
-gulp.task('style', compileCSS)
+gulp.task('clean:zip', async () => {
+  return clean(['dist/{assets,css,js}', 'dist/*.zip'])
+})
 
-gulp.task('build', gulp.series([compileCSS, webpackJS]))
+gulp.task('style', gulp.series([compileCSS, minifyCSS]))
+gulp.task('script', gulp.series([webpackJS, minifyJS]))
 
-gulp.task(
-  'bundle',
-  gulp.series([cleanDist, 'build', minifyCSS, minifyJS, versionJS])
-)
+gulp.task('build', gulp.series(['clean', gulp.parallel(['style', 'script'])]))
 
 gulp.task(
   'zip',
-  gulp.series(['bundle', assets, jsFolder, cssFolder, createZip])
+  gulp.series([
+    'clean:zip',
+    gulp.parallel([assets, jsFolder, cssFolder]),
+    createZip
+  ])
 )
 
 gulp.task('watch', () =>
   Promise.all([
-    gulp.watch(['packages/**/*.scss'], compileCSS),
-    gulp.watch(['packages/**/*.js'], webpackJS)
+    gulp.watch(['packages/**/*.scss'], gulp.series(['style'])),
+    gulp.watch(['packages/**/*.js'], gulp.series(['script']))
   ])
 )
 
 /**
- * The default task is to build everything, serve the docs and watch for changes
+ * The default task is to serve the docs and watch for changes
  */
-gulp.task(
-  'default',
-  gulp.series([
-    cleanDist,
-    'build',
-    'docs:build',
-    gulp.parallel(['docs:serve', 'docs:watch', 'watch'])
-  ])
-)
+gulp.task('default', gulp.parallel(['docs:serve', 'docs:watch', 'watch']))