+++ tests/backstop/engine_scripts/chromy/onReady.js
@@ -1,8 +1,0 @@
-const clickAndHoverHelper = require('./clickAndHoverHelper')
-
-module.exports = function (chromy, scenario) {
-  console.log(`SCENARIO > ${scenario.label}`)
-  clickAndHoverHelper(chromy, scenario)
-
-  // add more ready handlers here...
-}