+++ tests/backstop/engine_scripts/playwright/interceptImages.js
@@ -16,22 +16,16 @@ const fs = require('fs');
 const path = require('path');
 
 const IMAGE_URL_RE = /\.gif|\.jpg|\.png/i;
-const IMAGE_STUB_URL = path.resolve(__dirname, '../imageStub.jpg');
+const IMAGE_STUB_URL = path.resolve(__dirname, '../../imageStub.jpg');
 const IMAGE_DATA_BUFFER = fs.readFileSync(IMAGE_STUB_URL);
 const HEADERS_STUB = {};
 
 module.exports = async function (page, scenario) {
-  const intercept = async (request, targetUrl) => {
-    if (IMAGE_URL_RE.test(request.url())) {
-      await request.respond({
-        body: IMAGE_DATA_BUFFER,
-        headers: HEADERS_STUB,
-        status: 200
-      });
-    } else {
-      request.continue();
-    }
-  };
-  await page.setRequestInterception(true);
-  page.on('request', intercept);
+  page.route(IMAGE_URL_RE, route => {
+    route.fulfill({
+      body: IMAGE_DATA_BUFFER,
+      headers: HEADERS_STUB,
+      status: 200
+    });
+  });
 };